chore: cleanup refactor of PR
